fix(bluesky): sorted tests
